Moves Image Converter Code
Separates image converter from imagenexus into img_converter.py
