+++ imagenexus.py
@@ -4,6 +4,8 @@ import os
 from src.core.qr_generator import QRGenerator
 from src.core.pixelizer import Pixelize
 from src.core.frame_extractor import FrameExtractor
+from src.core.img_converter import ImgConverter
+
 from src.ui.ui_form import Ui_ImageNexus
 from src.utils.aboutDialog import aboutDialog
 from src.utils.version import appVersion
@@ -21,19 +23,21 @@ class ImageNexus(QMainWindow):
         self.ui = Ui_ImageNexus()
         self.ui.setupUi(self)
         self.setWindowTitle(f"ImageNexus v{version}")
-        self.pixelizer = Pixelize(self.ui)
-        self.qr_generator = QRGenerator(self.ui)
         self.frame_extractor = FrameExtractor(self.ui)
+        self.img_converter = ImgConverter(self.ui)
+        self.qr_generator = QRGenerator(self.ui)
+        self.pixelizer = Pixelize(self.ui)
+
         self.setup_connections()
         self.qr_generator.load_qr_templates()
 
 
     def setup_connections(self):
 
-        # Image Converter tab
-        self.ui.inputBrowse2.clicked.connect(self.select_input_file)
-        self.ui.outputBrowse2.clicked.connect(self.select_output_folder_converter)
-        self.ui.converter_button.clicked.connect(self.convert_file)
+        # # Image Converter tab
+        # self.ui.inputBrowse2.clicked.connect(self.select_input_file)
+        # self.ui.outputBrowse2.clicked.connect(self.select_output_folder_converter)
+        # self.ui.converter_button.clicked.connect(self.convert_file)
 
         # Batch Converter tab
         self.ui.inputBrowse3.clicked.connect(self.select_batch_input)
@@ -61,17 +65,6 @@ class ImageNexus(QMainWindow):
         self.about_dialog.show()
 
 
-    def select_input_file(self):
-        file_path, _ = QFileDialog.getOpenFileName(self, "Select Input File", "", "Image files (*.gif *.png *.jpg *.jpeg *.bmp *.tiff)")
-        if file_path:
-            self.ui.fileInput2.setText(file_path)
-            self.input_format = os.path.splitext(file_path)[1][1:].lower()
-
-    def select_output_folder_converter(self):
-        folder_path = QFileDialog.getExistingDirectory(self, "Select Output Folder")
-        if folder_path:
-            self.ui.outputFolder2.setText(folder_path)
-
     def select_batch_input(self):
         conversion_type = self.ui.conversionType.currentText()
         if conversion_type == "Files":
@@ -86,52 +79,6 @@ class ImageNexus(QMainWindow):
         if folder_path:
             self.ui.outputFolder3.setText(folder_path)
 
-
-
-    def convert_file(self):
-        input_path = self.ui.fileInput2.text()
-        output_folder = self.ui.outputFolder2.text()
-        output_format = self.ui.saveAsFormat_2.currentText().lower()
-
-        if not input_path or not output_folder:
-            QMessageBox.critical(self, "Error", "Please select both input file and output folder.")
-            return
-
-        input_filename = os.path.basename(input_path)
-        output_filename = os.path.splitext(input_filename)[0] + f".{output_format}"
-        output_path = os.path.join(output_folder, output_filename)
-
-        if not os.path.exists(output_folder):
-            try:
-                os.makedirs(output_folder)
-            except OSError as e:
-                QMessageBox.critical(self, "Error", f"Failed to create output folder: {e}")
-                return
-
-        if os.path.isfile(output_path):
-            overwrite = QMessageBox.question(self, "Overwrite Existing File",
-                "An existing file with the same name already exists.\n\nDo you want to overwrite it?",
-                QMessageBox.Yes | QMessageBox.No)
-            if overwrite == QMessageBox.No:
-                return
-
-        try:
-            with Image.open(input_path) as img:
-                if self.input_format == 'gif' and output_format != 'gif':
-                    img.seek(0)
-                    if output_format == 'jpeg':
-                        if img.mode == 'RGBA':
-                            img = img.convert('RGB')
-                    elif output_format == 'bmp':
-                        if img.mode != 'RGB':
-                            img = img.convert('RGB')
-                img.save(output_path, format=output_format.upper())
-
-            self.ui.statusbar.showMessage("Image converted successfully!")
-            QMessageBox.information(self, "Success", "Image converted successfully!")
-        except Exception as e:
-            self.ui.statusbar.showMessage(f"Error: {str(e)}")
-            QMessageBox.critical(self, "Error", f"An error occurred: {str(e)}")
 
     def convert_batch(self):
         input_paths = self.ui.fileInput3.text()
